food buy/sell functions working

+++ Budget/Budget.cpp
@@ -3,9 +3,9 @@
 
 using namespace std;
 
-Budget::Budget(int budget): m_budget(budget){}
+Budget::Budget(float budget): m_budget(budget){}
 
-void Budget::setBudget(int budget) {
+void Budget::setBudget(float budget) {
     if (budget < 0) {
         cout << "Vous ne pouvez pas avoir un budget negatif !" << endl;
         return;
@@ -13,9 +13,9 @@ void Budget::setBudget(int budget) {
     m_budget = budget;
 }
 
-int Budget::getBudget() { return m_budget; }
+float Budget::getBudget() { return m_budget; }
 
-void Budget::addBudget(int budget) {
+void Budget::addBudget(float budget) {
     if (budget < 0) {
         cout << "Vous ne pouvez pas ajouter une somme negative !" << endl;
         return;
@@ -23,7 +23,7 @@ void Budget::addBudget(int budget) {
     m_budget += budget;
 }
 
-void Budget::removeBudget(int budget) {
+void Budget::removeBudget(float budget) {
     if (budget > m_budget) {
         cout << "Vous ne pouvez pas depenser plus que votre budget !" << endl;
         return;
